Finished sendAll() implementation
