+++ test/integration/messaging.spec.ts
@@ -88,12 +88,14 @@ describe('admin.messaging', () => {
       });
   });
 
-  it.only('sendBatch()', () => {
-    const messages: admin.messaging.Message[] = [message];
-    return admin.messaging().sendBatch(messages, true)
-      .then((responses) => {
-        expect(responses.length).to.equal(messages.length);
-        responses.forEach((response) => {
+  it('sendAll()', () => {
+    const messages: admin.messaging.Message[] = [message, message, message];
+    return admin.messaging().sendAll(messages, true)
+      .then((response) => {
+        expect(response.responses.length).to.equal(messages.length);
+        expect(response.successCount).to.equal(messages.length);
+        expect(response.failureCount).to.equal(0);
+        response.responses.forEach((response) => {
           expect(response.success).to.be.true;
           expect(response.messageId).matches(/^projects\/.*\/messages\/.*$/);
         });
